refactored property validators into validators.py
